multimodal-assistant/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
pytorch-lightning>=2.2.0
transformers>=4.44.0
sentence-transformers>=2.7.0
accelerate>=0.33.0
optimum>=1.21.0
timm>=1.0.8
opencv-python>=4.9.0.80
Pillow>=10.3.0
numpy>=1.26.4
scipy>=1.12.0
scikit-learn>=1.4.2
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.5.2
python-dotenv>=1.0.1
loguru>=0.7.2
einops>=0.8.0
matplotlib>=3.8.4

# RAG & Vector Store
faiss-cpu>=1.8.0
pdfminer.six>=20240706

# Audio
soundfile>=0.12.1
# webrtcvad>=2.0.10
pyttsx3>=2.90

# Real-Time Computer Vision
mediapipe>=0.10.14
# face-recognition>=1.3.0  # Optional: requires CMake on Windows

# Development & Testing
pre-commit>=3.6.0
black>=24.4.2
ruff>=0.4.8
isort>=5.13.2
pytest>=8.1.1
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.27.0
pytest-mock>=3.11.1
aiohttp>=3.8.0
